+++ app/controllers/async/forum_controller.rb
@@ -2,7 +2,7 @@
 
 # Provides methods to login and register while starting a new forum topic.
 class Async::ForumController < Async::BaseController
-  before_action :set_chapter
+  before_action :set_topic
 
   def login
     super
@@ -13,12 +13,12 @@ class Async::ForumController < Async::BaseController
     end
   end
 
-  def new_topic
-  end
+  # def new_topic
+  # end
 
   private
 
-  def set_chapter
-    @chapter = Chapter.new
+  def set_topic
+    @topic = Topic.new
   end
 end