+++ app/controllers/replies_controller.rb
@@ -3,8 +3,8 @@
 # Allows users to reply to forum topics.
 class RepliesController < ApplicationController
   before_action :authenticate_user!, except: [:vote, :show]
-  before_action :find_chapter_or_redirect, except: [:preview, :show]
-  before_action :find_viewable_reply_or_redirect, only: [:show]
+  before_action :find_parent_or_redirect, only: [:create]
+  before_action :find_viewable_reply_or_redirect, only: [:show, :vote]
   before_action :find_editable_reply_or_redirect, only: [:edit, :update, :destroy]
 
   # # GET /replies
@@ -16,7 +16,13 @@ class RepliesController < ApplicationController
   def show
     respond_to do |format|
       format.html do
-        redirect_to page_chapter_path(@reply.chapter, page: @reply.page, anchor: @reply.anchor)
+        if @reply.parent.is_a?(Broadcast)
+          redirect_to blog_post_path(@reply.parent.url_hash.merge(page: @reply.page, anchor: @reply.anchor))
+        elsif @reply.parent.is_a?(Topic)
+          redirect_to page_topic_path(@reply.parent, page: @reply.page, anchor: @reply.anchor)
+        else
+          redirect_to root_path
+        end
       end
       format.js
     end
@@ -27,20 +33,17 @@ class RepliesController < ApplicationController
   end
 
   def vote
-    @reply = @chapter.replies.find_by_id params[:id]
-    redirect_without_reply
-    if current_user
-      @reply_user = @chapter.reply_users.where(
-        user_id: current_user.id, reply: @reply.id
-      ).first_or_create
-      case params[:vote]
-      when 'up'
-        @reply_user.up_vote!
-      when 'down'
-        @reply_user.down_vote!
-      else
-        @reply_user.remove_vote!
-      end
+    return unless current_user
+    @reply_user = @reply.parent.reply_users.where(
+      user_id: current_user.id, reply: @reply.id
+    ).first_or_create
+    case params[:vote]
+    when 'up'
+      @reply_user.up_vote!
+    when 'down'
+      @reply_user.down_vote!
+    else
+      @reply_user.remove_vote!
     end
   end
 
@@ -49,19 +52,18 @@ class RepliesController < ApplicationController
   #   @reply = Reply.new
   # end
 
-  # GET /replies/1/edit.js
-  def edit
-  end
+  # # GET /replies/1/edit.js
+  # def edit
+  # end
 
   # POST /replies
   def create
-    @reply = current_user.replies.where(chapter_id: @chapter.id)
-                         .new(reply_params)
+    @reply = @parent.replies.where(user_id: current_user.id).new(reply_params)
     if @reply.save
       @reply.create_notifications!
       @reply.compute_shadow_ban!
-      @reply.chapter.touch(:last_reply_at) unless @reply.user.shadow_banned?
-      current_user.read_chapter!(@chapter, @reply.id)
+      @reply.parent.touch(:last_reply_at) unless @reply.user.shadow_banned?
+      current_user.read_parent!(@parent, @reply.id)
       render :create
     else
       render :new
@@ -86,27 +88,25 @@ class RepliesController < ApplicationController
 
   private
 
-  def find_chapter_or_redirect
-    @chapter = Chapter.current.find_by_slug params[:chapter_id]
-    redirect_without_chapter
-  end
-
-  def redirect_without_chapter
-    empty_response_or_root_path(chapters_path) unless @chapter
+  def find_parent_or_redirect
+    @topic = Topic.current.find_by(slug: params[:topic_id])
+    @broadcast = Broadcast.current.published.find_by(slug: params[:broadcast_id])
+    @parent = @topic || @broadcast
+    empty_response_or_root_path unless @parent
   end
 
   def find_viewable_reply_or_redirect
-    @reply = Reply.current.find_by_id params[:id]
+    @reply = Reply.current.find_by(id: params[:id])
     redirect_without_reply
   end
 
   def find_editable_reply_or_redirect
-    @reply = current_user.editable_replies.where(chapter_id: @chapter.id).find_by_id params[:id]
+    @reply = current_user.editable_replies.find_by(id: params[:id])
     redirect_without_reply
   end
 
   def redirect_without_reply
-    empty_response_or_root_path(@chapter || chapters_path) unless @reply
+    empty_response_or_root_path unless @reply
   end
 
   def reply_params