+++ app/models/notification.rb
@@ -8,8 +8,7 @@ class Notification < ApplicationRecord
   # Model Relationships
   belongs_to :user
   belongs_to :broadcast
-  belongs_to :broadcast_comment
-  belongs_to :chapter
+  belongs_to :topic
   belongs_to :reply
 
   # Notification Methods