+++ test/controllers/admin/replies_controller_test.rb
@@ -2,7 +2,7 @@
 
 require 'test_helper'
 
-# Tests to assure that admins can review forum replies
+# Tests to assure that admins can review forum replies.
 class Admin::RepliesControllerTest < ActionController::TestCase
   test 'should get index' do
     login(users(:owner))