+++ test/controllers/members_controller_test.rb
@@ -5,7 +5,7 @@ require 'test_helper'
 class MembersControllerTest < ActionController::TestCase
   test 'should get index and redirect to forums' do
     get :index
-    assert_redirected_to chapters_path
+    assert_redirected_to topics_path
   end
 
   test 'should get show' do