+++ test/models/question_test.rb
@@ -3,5 +3,4 @@
 require 'test_helper'
 
 class QuestionTest < ActiveSupport::TestCase
-
 end